requirements.txt
# python==3.8
# conda install pytorch==1.12.0 torchvision==0.13.0 cudatoolkit=10.2 -c pytorch
torch==1.13
torchvision==0.14
scikit-learn==1.2.2
pandas==1.5.3
opencv-python
scipy==1.10.1
tqdm==4.65.0
matplotlib==3.7.1
pyyaml==6.0
hydra-core==1.3.2
torchsummary==1.5.1
hydra-optuna-sweeper==1.2.0
uvicorn
fastapi
scikit-image
python-multipart==0.0.5
optuna==2.10.1
chardet==5.2.0
jupyterlab==3.2.1
notebook 
pillow == 10.1.0
jupyter == 1.0.0
timm == 0.9.12
transformers==4.37.2